requirements.txt
numpy
numba
Pillow
soundfile
scipy